refactor: refactor main client component
